# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_api.ConferenceCup:
    properties:
      qf:
        description: QF[0]=1v4, QF[1]=2v3
        items:
          $ref: '#/definitions/internal_api.CupGame'
        type: array
      seeds:
        description: index 0 = seed 1, index 3 = wildcard
        items:
          type: string
        type: array
      sf:
        $ref: '#/definitions/internal_api.CupGame'
    type: object
  internal_api.ConferencePlayIn:
    properties:
      game1:
        $ref: '#/definitions/internal_api.PlayInGame'
      game2:
        $ref: '#/definitions/internal_api.PlayInGame'
      game3:
        $ref: '#/definitions/internal_api.PlayInGame'
      playoff_7:
        type: string
      playoff_8:
        type: string
    type: object
  internal_api.ConferencePlayoffBracket:
    properties:
      champion:
        example: BOS
        type: string
      r1:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeries'
        type: array
      r2:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeries'
        type: array
      r3:
        $ref: '#/definitions/internal_api.PlayoffSeries'
    type: object
  internal_api.CupBracket:
    properties:
      east:
        $ref: '#/definitions/internal_api.ConferenceCup'
      final:
        $ref: '#/definitions/internal_api.CupGame'
      west:
        $ref: '#/definitions/internal_api.ConferenceCup'
    type: object
  internal_api.CupGame:
    properties:
      away:
        type: string
      away_score:
        type: integer
      counted:
        type: boolean
      home:
        type: string
      home_score:
        type: integer
      winner:
        type: string
    type: object
  internal_api.DraftPick:
    properties:
      pick:
        example: 1
        type: integer
      seed:
        example: 1
        type: integer
      team_id:
        example: DET
        type: string
    type: object
  internal_api.ErrorResponse:
    properties:
      error:
        example: unknown team id
        type: string
    type: object
...
  internal_api.FIFAPenaltyKick:
    properties:
      kick_no:
        type: integer
      scored:
        type: boolean
      team:
        type: string
    type: object
  internal_api.FIFAPenaltyResult:
    properties:
      kicks:
        items:
          $ref: '#/definitions/internal_api.FIFAPenaltyKick'
        type: array
      team1_scored:
        type: integer
      team2_scored:
        type: integer
    type: object
  internal_api.FIFASimulateRequest:
    properties:
      extra_time:
        example: false
        type: boolean
      host_team_id:
        description: 'optional: team receiving +100 Elo crowd boost'
        example: USA
        type: string
      seed:
        example: 42
        type: integer
      team1_id:
        example: USA
        type: string
      team2_id:
        example: MEX
        type: string
    type: object
  internal_api.FIFASimulateResponse:
    properties:
      penalties:
        $ref: '#/definitions/internal_api.FIFAPenaltyResult'
      period:
        description: '"90" | "aet" | "pens"'
        type: string
      seed:
        type: integer
      team1_goals:
        type: integer
      team1_goals_aet:
        type: integer
      team1_id:
        type: string
      team1_name:
        type: string
      team2_goals:
        type: integer
      team2_goals_aet:
        type: integer
      team2_id:
        type: string
      team2_name:
        type: string
      winner:
        description: empty = draw (group stage)
        type: string
    type: object
...
  internal_api.GameEvent:
    properties:
      away_score:
        example: 98
        type: integer
      clock_sec:
        example: 45
        type: integer
      home_score:
        example: 105
        type: integer
      period:
        example: Q4
        type: string
      points:
        example: 2
        type: integer
      possession:
        example: 1
        type: integer
      team:
        example: home
        type: string
      text:
        example: Lakers make_2pt (2 pts)
        type: string
      type:
        example: make_2pt
        type: string
    type: object
  internal_api.GameState:
    properties:
      away_id:
        example: BOS
        type: string
      away_name:
        example: Celtics
        type: string
      away_score:
        example: 98
        type: integer
      clock_sec:
        example: 0
        type: integer
      home_id:
        example: LAL
        type: string
      home_name:
        example: Lakers
        type: string
      home_score:
        example: 105
        type: integer
      offense:
        description: omitted when status is final
        example: home
        type: string
      period:
        example: Q4
        type: string
      possession:
        example: 201
        type: integer
      quarter:
        example: 4
        type: integer
      status:
        example: final
        type: string
    type: object
...
  internal_api.PlayInGame:
    properties:
      away:
        type: string
      away_score:
        type: integer
      home:
        type: string
      home_score:
        type: integer
      loser:
        type: string
      winner:
        type: string
    type: object
  internal_api.PlayInTeams:
    properties:
      seed7:
        example: MIL
        type: string
      seed8:
        example: MIA
        type: string
      seed9:
        example: CHI
        type: string
      seed10:
        example: ATL
        type: string
    type: object
  internal_api.PlayoffBracketInput:
    properties:
      east_r1:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeriesInput'
        type: array
      east_r2:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeriesInput'
        type: array
      east_r3:
        $ref: '#/definitions/internal_api.PlayoffSeriesInput'
      finals:
        $ref: '#/definitions/internal_api.PlayoffSeriesInput'
      west_r1:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeriesInput'
        type: array
      west_r2:
        items:
          $ref: '#/definitions/internal_api.PlayoffSeriesInput'
        type: array
      west_r3:
        $ref: '#/definitions/internal_api.PlayoffSeriesInput'
    type: object
  internal_api.PlayoffGame:
    properties:
      away:
        example: MIA
        type: string
      away_score:
        example: 98
        type: integer
      game_num:
        example: 1
        type: integer
      home:
        example: BOS
        type: string
      home_score:
        example: 105
        type: integer
      winner:
        example: BOS
        type: string
    type: object
  internal_api.PlayoffSeries:
    properties:
      away_team:
        example: MIA
        type: string
      away_wins:
        example: 2
        type: integer
      games:
        items:
          $ref: '#/definitions/internal_api.PlayoffGame'
        type: array
      home_team:
        example: BOS
        type: string
      home_wins:
        example: 4
        type: integer
      winner:
        example: BOS
        type: string
    type: object
  internal_api.PlayoffSeriesInput:
    properties:
      away_team:
        example: MIA
        type: string
      away_wins:
        example: 1
        type: integer
      home_team:
        example: BOS
        type: string
      home_wins:
        example: 2
        type: integer
    type: object
  internal_api.PlayoffTeamInput:
    properties:
      season_l:
        example: 22
        type: integer
      season_w:
        example: 60
        type: integer
      seed:
        example: 1
        type: integer
      team_id:
        example: BOS
        type: string
    type: object
...
  internal_api.SimulateDraftLotteryRequest:
    properties:
      seed:
        example: 42
        type: integer
      teams:
        items:
          type: string
        type: array
    type: object
  internal_api.SimulateDraftLotteryResponse:
    properties:
      picks:
        items:
          $ref: '#/definitions/internal_api.DraftPick'
        type: array
      seed:
        example: 42
        type: integer
    type: object
  internal_api.SimulatePlayInRequest:
    properties:
      east:
        $ref: '#/definitions/internal_api.PlayInTeams'
      season:
        example: 2024-25
        type: string
      seed:
        example: 42
        type: integer
      west:
        $ref: '#/definitions/internal_api.PlayInTeams'
    type: object
  internal_api.SimulatePlayInResponse:
    properties:
      east:
        $ref: '#/definitions/internal_api.ConferencePlayIn'
      seed:
        type: integer
      west:
        $ref: '#/definitions/internal_api.ConferencePlayIn'
    type: object
  internal_api.SimulatePlayoffsRequest:
    properties:
      bracket:
        $ref: '#/definitions/internal_api.PlayoffBracketInput'
      east:
        items:
          $ref: '#/definitions/internal_api.PlayoffTeamInput'
        type: array
      season:
        example: 2024-25
        type: string
      seed:
        example: 42
        type: integer
      west:
        items:
          $ref: '#/definitions/internal_api.PlayoffTeamInput'
        type: array
    type: object
  internal_api.SimulatePlayoffsResponse:
    properties:
      champion:
        example: BOS
        type: string
      east:
        $ref: '#/definitions/internal_api.ConferencePlayoffBracket'
      finals:
        $ref: '#/definitions/internal_api.PlayoffSeries'
      seed:
        example: 42
        type: integer
      west:
        $ref: '#/definitions/internal_api.ConferencePlayoffBracket'
    type: object
  internal_api.SimulateRequest:
    properties:
      away_team_id:
        example: BOS
        type: string
      home_team_id:
        example: LAL
        type: string
      season:
        example: 2024-25
        type: string
      seed:
        example: 42
        type: integer
    type: object
  internal_api.SimulateResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/internal_api.GameEvent'
        type: array
      seed:
        example: 42
        type: integer
      state:
        $ref: '#/definitions/internal_api.GameState'
      truncated:
        example: false
        type: boolean
    type: object
  internal_api.SimulateSeasonRequest:
    properties:
      season:
        example: 2024-25
        type: string
      seed:
        example: 42
        type: integer
    type: object
  internal_api.SimulateSeasonResponse:
    properties:
      cup:
        $ref: '#/definitions/internal_api.CupBracket'
      east:
        items:
          $ref: '#/definitions/internal_api.TeamSeasonStat'
        type: array
      season:
        type: string
      seed:
        type: integer
      west:
        items:
          $ref: '#/definitions/internal_api.TeamSeasonStat'
        type: array
    type: object
...
  internal_api.TeamSeasonStat:
    properties:
      away_record:
        example: 24-17
        type: string
      conf_record:
        example: 32-20
        type: string
      conference:
        example: west
        type: string
      diff:
        example: 4.4
        type: number
      div_record:
        example: 10-4
        type: string
      division:
        example: pacific
        type: string
      home_record:
        example: 30-11
        type: string
      l:
        example: 28
        type: integer
      last_10:
        example: 7-3
        type: string
      oppg:
        example: 109.8
        type: number
      ppg:
        example: 114.2
        type: number
      streak:
        example: W3
        type: string
      team_id:
        example: LAL
        type: string
      team_name:
        example: Lakers
        type: string
      w:
        example: 54
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Possession-by-possession NBA game simulator API.
  title: Sim1.0 API
  version: "1.0"
paths:
  /fifa/simulate:
    post:
      consumes:
      - application/json
      description: Uses a Poisson goals model with Dixon-Coles correction. ExtraTime=true
        adds AET and penalties if the match is drawn.
      parameters:
      - description: Home/away team ids, optional seed, extra_time flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.FIFASimulateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.FIFASimulateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Simulate a FIFA World Cup match
      tags:
      - fifa
...
  /nba/simulate:
    post:
      consumes:
      - application/json
      description: Runs possessions until the game clock ends. Same seed produces
        the same game.
      parameters:
      - description: Home/away team ids and optional seed
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.SimulateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimulateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Simulate a full game
      tags:
      - simulate
  /nba/simulate/draft-lottery:
    post:
      consumes:
      - application/json
      description: Runs the NBA draft lottery for 14 teams using official ball-combination
        odds. Picks 1-4 are drawn by weighted lottery; picks 5-14 go to remaining
        teams in seed order.
      parameters:
      - description: 14 team IDs in lottery-seed order (index 0 = worst record)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.SimulateDraftLotteryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimulateDraftLotteryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Simulate the NBA draft lottery
      tags:
      - simulate
  /nba/simulate/playin:
    post:
      consumes:
      - application/json
      description: 'Runs all 6 play-in games (3 per conference). Game 1: 7 hosts 8
        — winner = 7 seed. Game 2: 9 hosts 10. Game 3: loser of G1 hosts winner of
        G2 — winner = 8 seed.'
      parameters:
      - description: Play-in team IDs for each conference
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.SimulatePlayInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimulatePlayInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Simulate the NBA play-in tournament
      tags:
      - simulate
  /nba/simulate/playoffs:
    post:
      consumes:
      - application/json
      description: Simulates all playoff series from the current bracket state to
        the champion. Each conference seeds 1-8; matchups follow standard NBA seeding
        (1v8, 4v5, 3v6, 2v7). Home court follows HHAAAHA schedule. Finals home court
        goes to the team with the better regular-season record (coin flip on a tie).
        Partially completed series are resumed from their current wins.
      parameters:
      - description: 16 playoff teams (8 east, 8 west) with seeds and records; optional
          partial bracket state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.SimulatePlayoffsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimulatePlayoffsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Simulate the NBA playoffs
      tags:
      - simulate
  /nba/simulate/season:
    post:
      consumes:
      - application/json
      description: Runs every game in the 2025-26 schedule and returns standings with
        W, L, streak, last-10, home/away records, PPG, OPPG, and DIFF.
      parameters:
      - description: Optional seed for reproducibility
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_api.SimulateSeasonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimulateSeasonResponse'
      summary: Simulate a full NBA regular season
      tags:
      - simulate
...
swagger: "2.0"